chore: refactored jobValidate method into separate module
